NXP-3152 - use another cache directory in case of access denied, test existence before rights

+++ nuxeo-core-convert/src/main/java/org/nuxeo/ecm/core/convert/extension/GlobalConfigDescriptor.java
@@ -93,16 +93,16 @@ public class GlobalConfigDescriptor implements Serializable {
     public String getCachingDirectory() {
         if (cachingDirectory == null) {
             File cacheFile=new File(System.getProperty("java.io.tmpdir"),CACHING_DIRECTORY);
-            if (!cacheFile.canWrite()) {
-                // change directory to avoid FileNotFoundException (permission denied)
+            if (cacheFile.exists() && !cacheFile.canWrite()) {
+                log.debug("change directory to avoid FileNotFoundException (permission denied)");
                 try {
                     cacheFile=File.createTempFile(CACHING_DIRECTORY, null, cacheFile.getParentFile());
                     cacheFile.delete();
-                    cacheFile.mkdirs();
                 } catch (IOException e) {
                     log.error("Could not create caching directory",e);
                 }
             }
+            cacheFile.mkdirs();
             cachingDirectory = cacheFile.getPath();
         }
         return cachingDirectory;
